Create mapLimit
In this question, you need to implement a custom mapLimit function that takes 4 arguments.

Arguments
inputs: An array of inputs
limit: The maximum number of operations at any given time.
iterateeFn: The async function that should be called with each input to generate the corresponding output. It will have two arguments:
input: The input being processed
callback: A function that will be called when the input is finished processing. It will be provided with one argument, the processed output.
callback: A function that should be called with the array of outputs once all inputs have been processed.
At any given point, your program can make max 2 calls i.e. at any given point your program can process 1, 2 or 2, 3 or so on user ids.